+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/camera/RecenterButtonComponentContractImpl.kt
@@ -5,13 +5,10 @@ import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.ui.app.internal.Store
 import com.mapbox.navigation.ui.app.internal.camera.CameraAction
 import com.mapbox.navigation.ui.app.internal.camera.TargetCameraMode
-import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.maps.internal.ui.RecenterButtonComponentContract
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.stateIn
 
 @ExperimentalPreviewMapboxNavigationAPI
 internal class RecenterButtonComponentContractImpl(
@@ -19,20 +16,12 @@ internal class RecenterButtonComponentContractImpl(
     private val store: Store
 ) : RecenterButtonComponentContract {
 
-    override val isVisible: StateFlow<Boolean> = combine(
-        store.select { it.camera.cameraMode },
-        store.select { it.navigation },
-        transform = ::isVisible
-    ).stateIn(
-        scope,
-        SharingStarted.Eagerly,
-        isVisible(store.state.value.camera.cameraMode, store.state.value.navigation)
-    )
+    override val isVisible: StateFlow<Boolean> =
+        store.slice(scope = scope, started = SharingStarted.Eagerly) { state ->
+            state.camera.cameraMode == TargetCameraMode.Idle
+        }
 
     override fun onClick(view: View) {
         store.dispatch(CameraAction.SetCameraMode(TargetCameraMode.Following))
     }
-
-    private fun isVisible(cameraMode: TargetCameraMode, navigationState: NavigationState) =
-        cameraMode == TargetCameraMode.Idle && navigationState != NavigationState.RoutePreview
 }